Update bike bus data from Jotform/WebhookDB (#20)
Instead of updating `buses.yml`, users can submit a Jotform form with the necessary data.

The form uses a shared secret as one of the inputs so that random users cannot modify this website.

The form submission sends a webhookdb to WebhookDB, which schematizes the payload and puts it into
a Postgres database.

The build process grabs the latest row for each school, and uses that data instead of what's in `buses.yml`
(whatever is in `buses.yml` is used as a fallback).

Images on the Jotform are downloaded and re-hosted in this repo (as compressed webp images).
Note that the images are checked in, because we may need to delete Jotform reponses,
which would delete the uploaded images.