Register new static files with the build
If the output file did not exist when the build started,
we need to explicitly add it as a static file,
so it's copied to the _site build folder.
Otherwise, the file is placed in the (source) assets directory,
but not copied over to the build folder,
so doesn't get included in production.